Improve chat UI (wip)
